FAD-6046 linting fixes
