+++ src/components/navigation/Footer.js
@@ -6,10 +6,12 @@ import styles from './Navigation.module.scss';
 
 export class Footer extends Component {
   render() {
+    const { logout } = this.props;
+
     return (
       <ul className={styles.footer}>
         <li><a className={styles.link} href='https://www.sparkpost.com/docs/' target='_blank' rel='noopener noreferrer'>Help &amp; API</a></li>
-        <li><a className={styles.link} onClick={() => this.props.logout()}>Log Out</a></li>
+        <li><a className={styles.link} onClick={logout}>Log Out</a></li>
       </ul>
     );
   }
