+++ src/config/routes.js
@@ -92,7 +92,7 @@ const routes = [
     path: '/dashboard',
     component: DashboardPage,
     layout: App,
-    condition: () => config.splashPage === 'dashboard' // want to hide if not a splash page https://jira.int.messagesystems.com/browse/FAD-6046
+    condition: composeConditions(hasGrants('api_keys/manage', 'templates/modify', 'sending_domains/manage'), () => config.splashPage === '/dashboard') // want to hide if not a splash page https://jira.int.messagesystems.com/browse/FAD-6046
   },
   {
     path: '/reports',
@@ -141,19 +141,19 @@ const routes = [
   {
     path: '/account/subaccounts',
     component: subaccounts.ListPage,
-    condition: hasGrants('subaccount/manage'),
+    condition: hasGrants('subaccount/manage', 'api_keys/manage', 'sending_domains/manage'),
     layout: App
   },
   {
     path: '/account/subaccounts/create',
     component: subaccounts.CreatePage,
-    condition: hasGrants('subaccount/manage'),
+    condition: hasGrants('subaccount/manage', 'api_keys/manage', 'sending_domains/manage'),
     layout: App
   },
   {
     path: '/account/subaccounts/:id',
     component: subaccounts.DetailsPage,
-    condition: hasGrants('subaccount/manage'),
+    condition: hasGrants('subaccount/manage', 'api_keys/manage', 'sending_domains/manage'),
     layout: App,
     exact: false
   },
